+++ packages/node-backend/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  ...require('@goatlab/eslint/eslint-server.js'),
+  parserOptions: {
+    tsconfigRootDir: __dirname,
+    project: './tsconfig.json'
+  }
+}